# Web.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "EventFeedback.Web.Api"
  },
  "host": "localhost:6003",
  "schemes": [ "http" ],
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": [ "Events" ],
        "operationId": "Events_Get",
        "consumes": [ ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Event" }
            }
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [ "Events" ],
        "operationId": "Events_Post",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "entity",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Event" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Event" }
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "tags": [ "Events" ],
        "operationId": "Events_GetById",
        "consumes": [ ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Event" }
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [ "Events" ],
        "operationId": "Events_Put",
        "consumes": [ "application/json", "text/json", "application/x-www-form-urlencoded" ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "entity",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Event" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Event" }
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [ "Events" ],
        "operationId": "Events_Delete",
        "consumes": [ ],
        "produces": [ "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Object" }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "Event": {
      "required": [ "title" ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "active": { "type": "string" },
        "createDate": { "type": "string" },
        "createdBy": { "type": "string" },
        "modifyDate": { "type": "string" },
        "modifiedBy": { "type": "string" },
        "deleted": { "type": "string" },
        "deleteDate": { "type": "string" },
        "deletedBy": { "type": "string" },
        "title": { "type": "string" },
        "description": { "type": "string" },
        "key": { "type": "string" },
        "link": { "type": "string" },
        "feedbackAllowed": { "type": "string" },
        "feedbackDefinitionId": { "type": "string" },
        "feedbackDefinition": { "$ref": "FeedbackDefinition" },
        "startDate": { "type": "string" },
        "endDate": { "type": "string" },
        "location": { "type": "string" },
        "organization": { "type": "string" },
        "activeFromDate": { "type": "string" },
        "activeTillDate": { "type": "string" },
        "tags": {
          "type": "array",
          "items": { "type": "string" }
        },
        "tagList": { "type": "string" },
        "feedbacks": {
          "type": "array",
          "items": { "$ref": "Feedback" }
        },
        "sessions": {
          "type": "array",
          "items": { "$ref": "Session" }
        }
      }
    },
...
    "Object": {
      "type": "object",
      "properties": { }
    },
...
  }
}